pyproject.toml
[project]
name = "manimations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "manim>=0.19.0",
    "python-dotenv>=1.1.0",
]
